--- src/tests/Unit/profile.test.js
@@ -1,1 +1,0 @@
-//todo: write tests
