+++ tests/Unit/profile.test.js
@@ -1,0 +1,2 @@
+//todo: write tests
+test("silence", () => {});
